// src/models/book.ts
// @ts-ignore
import client from '../database';

export type Book = {
  id?: number; //optional because it will be generated by the database
  title: string;
  author: string;
  totalPages: number;
  summary: string;
};

export class BookStore {
  async index(): Promise<Book[]> {
    try {
      const conn = await client.connect();
      const sql = 'SELECT id, title, author, total_pages as "totalPages", summary FROM books';

      const result = await conn.query(sql);
      conn.release();

      return result.rows;
    } catch (err) {
      throw new Error(`Could not get books. Error: ${err}`);
    }
  }

  async show(id: number): Promise<Book> {
    try {
      const sql = 'SELECT id, title, author, total_pages as "totalPages", summary FROM books WHERE id=($1)';
      const conn = await client.connect();

      const result = await conn.query(sql, [id]);
      conn.release();

      return result.rows[0];
    } catch (err) {
      throw new Error(`Could not find book ${id}. Error: ${err}`);
    }
  }

  async create(b: Book): Promise<Book> {
    try {
      const sql =
        'INSERT INTO books (title, author, total_pages, summary) VALUES($1, $2, $3, $4) RETURNING id, title, author, total_pages as "totalPages", summary';

      const conn = await client.connect();
      const result = await conn.query(sql, [b.title, b.author, b.totalPages, b.summary]);

      const book = result.rows[0];
      conn.release();

      return book;
    } catch (err) {
      throw new Error(`Could not add new book ${b.title}. Error: ${err}`);
    }
  }

  async delete(id: number): Promise<Book> {
    try {
      const sql =
        'DELETE FROM books WHERE id=($1) RETURNING id, title, author, total_pages as "totalPages", summary';

      const conn = await client.connect();
      const result = await conn.query(sql, [id]);

      const book = result.rows[0];
      conn.release();

      return book;
    } catch (err) {
      throw new Error(`Could not delete book ${id}. Error: ${err}`);
    }
  }

  async update(id: number, b: Book): Promise<Book> {
  try {
    const sql = `
      UPDATE books
      SET title = ($1), author = ($2), total_pages = ($3), summary = ($4)
      WHERE id = ($5)
      RETURNING id, title, author, total_pages as "totalPages", summary
    `;

    const conn = await client.connect();

    const result = await conn.query(sql, [
      b.title,
      b.author,
      b.totalPages,
      b.summary,
      id
    ]);

    const book = result.rows[0];
    conn.release();

    return book;
  } catch (err) {
    throw new Error(`Could not update book ${id}. Error: ${err}`);
  }
 }
}